move do_process_tracking_file
